Changes to get a clean build
