+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/model/ccd/wrapper/SendOrderWrapper.java
@@ -1,6 +1,5 @@
 package uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.wrapper;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
@@ -9,7 +8,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseDocument;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.DynamicMultiSelectList;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.DynamicMultiSelectListElement;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.SendOrderEventPostStateOption;
 
 import java.util.List;
@@ -30,21 +28,4 @@ public class SendOrderWrapper {
     private List<OrderToShareCollection> ordersToShareCollection;
 
     private SendOrderEventPostStateOption sendOrderPostStateOption;
-
-    private DynamicMultiSelectList partiesOnCase;
-
-    @JsonIgnore
-    public List<String> getSelectedParties() {
-        DynamicMultiSelectList parties = this.getPartiesOnCase();
-        return this.getSelectedParties(parties);
-    }
-
-    @JsonIgnore
-    public List<String> getSelectedParties(DynamicMultiSelectList parties) {
-        if (parties == null) {
-            return List.of();
-        }
-        return parties.getValue().stream().map(DynamicMultiSelectListElement::getCode).toList();
-    }
-
 }
