+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/GeneralOrderService.java
@@ -301,7 +301,7 @@ public class GeneralOrderService {
 
     public List<String> getParties(FinremCaseDetails caseDetails) {
         FinremCaseData data = caseDetails.getData();
-        DynamicMultiSelectList parties = data.getSendOrderWrapper().getPartiesOnCase();
+        DynamicMultiSelectList parties = data.getPartiesOnCase();
         return parties.getValue().stream().map(DynamicMultiSelectListElement::getCode).toList();
     }
