+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/SendOrderContestedAboutToSubmitHandlerTest.java
@@ -140,7 +140,7 @@ class SendOrderContestedAboutToSubmitHandlerTest {
 
         FinremCaseData caseData = response.getData();
 
-        assertNull(caseData.getSendOrderWrapper().getPartiesOnCase());
+        assertNull(caseData.getPartiesOnCase());
 
         verify(generalOrderService).getParties(callbackRequest.getCaseDetails());
         verify(generalOrderService).hearingOrdersToShare(callbackRequest.getCaseDetails(), null);
@@ -167,7 +167,7 @@ class SendOrderContestedAboutToSubmitHandlerTest {
         FinremCallbackRequest callbackRequest = buildCallbackRequest();
         FinremCaseDetails caseDetails = callbackRequest.getCaseDetails();
         FinremCaseData data = caseDetails.getData();
-        data.getSendOrderWrapper().setPartiesOnCase(new DynamicMultiSelectList());
+        data.setPartiesOnCase(new DynamicMultiSelectList());
 
         DynamicMultiSelectList selectedDocs = DynamicMultiSelectList.builder()
             .value(List.of(DynamicMultiSelectListElement.builder()
@@ -201,7 +201,7 @@ class SendOrderContestedAboutToSubmitHandlerTest {
             = handler.handle(callbackRequest, AUTH_TOKEN);
 
         FinremCaseData caseData = response.getData();
-        assertNull(caseData.getSendOrderWrapper().getPartiesOnCase().getValue());
+        assertNull(caseData.getPartiesOnCase().getValue());
         assertEquals(1, caseData.getFinalOrderCollection().size());
         assertNull(caseData.getOrderWrapper().getIntv1OrderCollection());
         assertNull(caseData.getOrderWrapper().getAppOrderCollection());
@@ -216,7 +216,7 @@ class SendOrderContestedAboutToSubmitHandlerTest {
         FinremCallbackRequest callbackRequest = buildCallbackRequest();
         FinremCaseDetails caseDetails = callbackRequest.getCaseDetails();
         FinremCaseData data = caseDetails.getData();
-        data.getSendOrderWrapper().setPartiesOnCase(new DynamicMultiSelectList());
+        data.setPartiesOnCase(new DynamicMultiSelectList());
 
         DynamicMultiSelectList selectedDocs = DynamicMultiSelectList.builder()
             .value(List.of(DynamicMultiSelectListElement.builder()
@@ -251,7 +251,7 @@ class SendOrderContestedAboutToSubmitHandlerTest {
             = handler.handle(callbackRequest, AUTH_TOKEN);
 
         FinremCaseData caseData = response.getData();
-        assertNull(caseData.getSendOrderWrapper().getPartiesOnCase().getValue());
+        assertNull(caseData.getPartiesOnCase().getValue());
         assertEquals(1, caseData.getFinalOrderCollection().size());
         assertNull(caseData.getOrderWrapper().getIntv1OrderCollection());
         assertNull(caseData.getOrderWrapper().getAppOrderCollection());
@@ -268,7 +268,7 @@ class SendOrderContestedAboutToSubmitHandlerTest {
         FinremCallbackRequest callbackRequest = buildCallbackRequest();
         FinremCaseDetails caseDetails = callbackRequest.getCaseDetails();
         FinremCaseData data = caseDetails.getData();
-        data.getSendOrderWrapper().setPartiesOnCase(getParties());
+        data.setPartiesOnCase(getParties());
 
         DynamicMultiSelectList selectedDocs = DynamicMultiSelectList.builder()
             .value(List.of(DynamicMultiSelectListElement.builder()
@@ -306,7 +306,7 @@ class SendOrderContestedAboutToSubmitHandlerTest {
             = handler.handle(callbackRequest, AUTH_TOKEN);
 
         FinremCaseData caseData = response.getData();
-        assertEquals(12, caseData.getSendOrderWrapper().getPartiesOnCase().getValue().size(), "selected parties on case");
+        assertEquals(12, caseData.getPartiesOnCase().getValue().size(), "selected parties on case");
         assertEquals(1, caseData.getFinalOrderCollection().size());
         assertNull(caseData.getOrderWrapper().getIntv1OrderCollection());
         assertEquals(1, caseData.getOrderWrapper().getIntv1OrderCollections().size());
@@ -326,7 +326,7 @@ class SendOrderContestedAboutToSubmitHandlerTest {
             = handler.handle(callbackRequest, AUTH_TOKEN);
 
         caseData = response.getData();
-        assertEquals(12, caseData.getSendOrderWrapper().getPartiesOnCase().getValue().size(), "selected parties on case");
+        assertEquals(12, caseData.getPartiesOnCase().getValue().size(), "selected parties on case");
         assertEquals(1, caseData.getFinalOrderCollection().size());
         assertNull(caseData.getOrderWrapper().getIntv1OrderCollection());
         assertEquals(2, caseData.getOrderWrapper().getIntv1OrderCollections().size());
@@ -365,7 +365,7 @@ class SendOrderContestedAboutToSubmitHandlerTest {
         FinremCallbackRequest callbackRequest = buildCallbackRequest();
         FinremCaseDetails caseDetails = callbackRequest.getCaseDetails();
         FinremCaseData data = caseDetails.getData();
-        data.getSendOrderWrapper().setPartiesOnCase(getParties());
+        data.setPartiesOnCase(getParties());
         List<DirectionOrderCollection> orderList = new ArrayList<>();
         DirectionOrderCollection order = DirectionOrderCollection.builder().value(DirectionOrder.builder()
             .uploadDraftDocument(caseDocument("http://abc/docurl", "abc.pdf", "http://abc/binaryurl"))
@@ -398,7 +398,7 @@ class SendOrderContestedAboutToSubmitHandlerTest {
             = handler.handle(callbackRequest, AUTH_TOKEN);
 
         FinremCaseData caseData = response.getData();
-        assertEquals(12, caseData.getSendOrderWrapper().getPartiesOnCase().getValue().size());
+        assertEquals(12, caseData.getPartiesOnCase().getValue().size());
         assertNull(caseData.getOrderWrapper().getIntv1OrderCollection());
         assertEquals(2, caseData.getFinalOrderCollection().size());
 
@@ -412,7 +412,7 @@ class SendOrderContestedAboutToSubmitHandlerTest {
         FinremCaseDetails caseDetails = callbackRequest.getCaseDetails();
         FinremCaseData data = caseDetails.getData();
         data.setOrderApprovedCoverLetter(caseDocument("coversheet", "coversheet"));
-        data.getSendOrderWrapper().setPartiesOnCase(getParties());
+        data.setPartiesOnCase(getParties());
         List<DirectionOrderCollection> orderList = new ArrayList<>();
         CaseDocument caseDocument = caseDocument("docurl", "abc.pdf", "binaryurl");
         DirectionOrderCollection order = DirectionOrderCollection.builder().value(DirectionOrder.builder()
@@ -443,7 +443,7 @@ class SendOrderContestedAboutToSubmitHandlerTest {
             = handler.handle(callbackRequest, AUTH_TOKEN);
 
         FinremCaseData caseData = response.getData();
-        assertEquals(12, caseData.getSendOrderWrapper().getPartiesOnCase().getValue().size());
+        assertEquals(12, caseData.getPartiesOnCase().getValue().size());
         assertNull(caseData.getOrderWrapper().getIntv1OrderCollection());
         assertEquals(1, caseData.getOrderWrapper().getIntv1OrderCollections().size());
         assertEquals(1, caseData.getFinalOrderCollection().size());
@@ -461,7 +461,7 @@ class SendOrderContestedAboutToSubmitHandlerTest {
 
         String additionalHearingDocumentFilename = "AdditionalHearingDocument.pdf";
 
-        data.getSendOrderWrapper().setPartiesOnCase(getParties());
+        data.setPartiesOnCase(getParties());
         ApproveOrder additionalHearingOrder = ApproveOrder.builder().orderReceivedAt(
                 LocalDateTime.of(LocalDate.of(2019, 12, 31),
                     LocalTime.of(6, 30, 45)))
@@ -595,7 +595,7 @@ class SendOrderContestedAboutToSubmitHandlerTest {
         FinremCallbackRequest callbackRequest = buildCallbackRequest();
         FinremCaseDetails caseDetails = callbackRequest.getCaseDetails();
         FinremCaseData data = caseDetails.getData();
-        data.getSendOrderWrapper().setPartiesOnCase(getParties());
+        data.setPartiesOnCase(getParties());
         List<DirectionOrderCollection> orderList = new ArrayList<>();
         DirectionOrderCollection order = DirectionOrderCollection.builder().value(DirectionOrder.builder()
             .uploadDraftDocument(caseDocument("http://abc/docurl", "abc.pdf", "http://abc/binaryurl"))
